refactor(security_entity_store): share 500-retry helper and retry entity store install
- Extract kibanaoapi.RetryCreateOnServerError shared by entity, entity-link,
  and entity-store install create paths (removes duplication)
- Apply 500-retry to entity store Install (Create + merge-on-install Update),
  since the isolation cleanup makes back-to-back uninstall/install races likely
- CleanupEntityStore guards on TF_ACC directly (no t.Skip inside t.Cleanup)
- waitForUninstall does an immediate first check before polling
- Add REQ-WAIT-003 spec + task 3.5
